Moving crypter classes to pinot-spi
