+++ pinot-spi/src/main/java/org/apache/pinot/spi/crypt/PinotCrypterFactory.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.core.crypt;
+package org.apache.pinot.spi.crypt;
 
 import java.util.HashMap;
 import java.util.Iterator;
